refactor(formatters): move function
